LoadNextLevel + All Levelhandler logic done
LoadNextLevel function complete that change that changes an index that keeps track of the level even if you close the game. Also now there is a new UI part that shows the current level. LevelHandler logic script is done in theory but there might still be some small changes here and there.