Implemented Scores system
Implemented in GameLogic the score sistem that keeps track of the combo you are on and gives you a multiplier if you keep matching before the time left runs out. There is sill needed a ui element like a bar to transmit how much time is left before the combo is lost.